# src/api/openapi.yaml
openapi: 3.1.0
info:
  title: SAVED4 OpenAPI Spec
  description: API specification for SAVED4
  version: 0.1.0
  x-fiddle-import-file: true
paths:
  /companies:
    post:
      operationId: addCompany
      summary: Add company
      responses:
        "201":
          description: The request was successful, and a new resource was created.
          content:
            application/json:
              schema:
                type: integer
                description: username of the newly created company user
                title: AddCompanyCreated
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalServerErrror"
      parameters: []
      tags:
        - companies
      security: []
      description: Creates a new company account in the system. Appropriate error
        responses are returned if creation fails.
      requestBody:
        required: true
        description: JSON object containing the company account details.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Company"
    get:
      operationId: getAllCompanies
      summary: Get all companies
      responses:
        "200":
          description: The request was successful, and the server has returned the
            requested resource in the response body.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CompanySummary"
                title: GetAllCompaniesOk
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerErrror"
      parameters:
        - in: query
          name: limit
          required: false
          style: form
          description: Maximum number of users to return,
          explode: true
          schema:
            type: integer
          allowReserved: false
      description: Retrieves an unordered list of company users from the system.
      tags:
        - companies
      security: []
  /companies/{username}:
    get:
      operationId: getCompany
      summary: Get company
      parameters:
        - $ref: "#/components/parameters/username"
      responses:
        "200":
          description: The request was successful, and the server has returned the
            requested resource in the response body.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanySummary"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerErrror"
      description: "Retrieves a company resource from the system by its username. If
        the company does not exist, a 404 response is returned. "
      tags:
        - companies
      security: []
    delete:
      operationId: deleteCompany
      summary: Delete company
      parameters:
        - $ref: "#/components/parameters/username"
      responses:
        "204":
          description: The request was successful, but there is no content to return in
            the response.
          content:
            application/json:
              schema:
                type: "null"
                title: DeleteCompanyNoContent
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerErrror"
      description: Deletes the company identified by the specified username.
      tags:
        - companies
      security: []
    patch:
      operationId: updateAnExistingCompany
      summary: Update an existing company
      parameters:
        - $ref: "#/components/parameters/username"
      responses:
        "200":
          description: The request was successful, and the server has returned the
            requested resource in the response body.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanySummary"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalServerErrror"
      description: Updates the details of an existing company identified by the
        username in the path.
      tags:
        - companies
      security: []
      requestBody:
        required: true
        description: A JSON object containing the fields to be updated.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanyPartialUpdate"
...
  /orders/company/{username}:
    get:
      operationId: getCompanyOrders
      summary: Get company orders
      description: Retrieves all the orders made by a company, sorted by newest first.
      parameters:
        - $ref: "#/components/parameters/username"
      responses:
        "200":
          description: The request was successful, and the server has returned the
            requested resource in the response body.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
                title: GetCompanyOrdersOk
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerErrror"
      tags:
        - orders
      security: []
...
  /chats/{chat_id}:
    get:
      summary: Get chat
      operationId: getChat
      description: Returns a chat identified by chat id.
      tags:
        - chats
      security: []
      parameters:
        - $ref: "#/components/parameters/chat_id"
      responses:
        "200":
          description: The request was successful, and the server has returned the
            requested resource in the response body.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chat"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerErrror"
  /chats:
    get:
      summary: List chat
      operationId: listChat
      description: Returns a list of all chats accessible to the client, ordered by
        most recent activity.
      tags:
        - chats
      security: []
      responses:
        "200":
          description: The request was successful, and the server has returned the
            requested resource in the response body.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                title: ListChatOk
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerErrror"
      parameters:
        - in: query
          name: limit
          required: false
          style: form
          description: Maximum number of chats to return
          explode: true
          schema:
            type: integer
          allowReserved: false
...
  /containers/{container_id}:
    get:
      summary: Get container
      operationId: getContainer
      description: Returns a container identified by container id.
      tags:
        - containers
      security: []
      parameters:
        - $ref: "#/components/parameters/container_id"
      responses:
        "200":
          description: The request was successful, and the server has returned the
            requested resource in the response body.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Container"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerErrror"
    delete:
      summary: Delete container
      operationId: deleteContainer
      description: Deletes a container.
      tags:
        - containers
      security: []
      parameters:
        - $ref: "#/components/parameters/container_id"
      responses:
        "204":
          description: The request was successful, but there is no content to return in
            the response.
          content:
            application/json:
              schema:
                type: "null"
                title: DeleteContainerNoContent
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerErrror"
  /containers:
    post:
      summary: Create container
      operationId: createContainer
      description: Creates a new record of type Container.
      tags:
        - containers
      security: []
      responses:
        "201":
          description: The request was successful, and a new resource was created.
          content:
            application/json:
              schema:
                type: integer
                description: id of the newly created container
                title: CreateContainerCreated
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalServerErrror"
      requestBody:
        required: true
        description: ""
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContainerCreate"
    get:
      summary: List container
      operationId: listContainer
      description: Returns a list of all containers.
      tags:
        - containers
      security: []
      responses:
        "200":
          description: The request was successful, and the server has returned the
            requested resource in the response body.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Container"
                title: ListContainerOk
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerErrror"
      parameters:
        - in: query
          name: limit
          required: false
          style: form
          description: Maximum number of chats to return
          explode: true
          schema:
            type: integer
          allowReserved: false
...
components:
  parameters:
    username:
      name: username
      in: path
      required: true
      schema:
        type: string
...
    chat_id:
      name: chat_id
      in: path
      required: true
      schema:
        type: integer
...
    container_id:
      name: container_id
      in: path
      required: true
      schema:
        type: integer
...
  responses:
    BadRequest:
      description: The server could not understand the request due to invalid syntax.
        The client should modify the request and try again.
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
              errors:
                type: array
                items:
                  type: object
                  properties:
                    message:
                      type: string
                  required:
                    - message
            required:
              - message
    Unauthorized:
      description: Authentication is required to access the requested resource. The
        client must include the appropriate credentials.
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
            required:
              - message
    Forbidden:
      description: The server understood the request, but refuses to authorize it.
        Ensure the client has appropriate permissions.
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
            required:
              - message
    InternalServerErrror:
      description: The server encountered an unexpected condition that prevented it
        from fulfilling the request. Report the issue to the support team if it
        persists.
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
            required:
              - message
    NotFound:
      description: The server cannot find the requested resource. The endpoint may be
        invalid or the resource may no longer exist.
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
            required:
              - message
...
    Conflict:
      description: The request could not be completed due to a conflict with the
        current state of the resource. Resolve the conflict and try again.
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
            required:
              - message
...
  securitySchemes:
    BearerJwt:
      type: http
      scheme: Bearer
      bearerFormat: Json Web Token (JWT)
      description: "Jwt Auth: Authenticated requests contain a valid Json Web Token
        (JWT) as part of the `Authorization: Bearer <JWT>` header."
  schemas:
    chat_id:
      type: integer
      description: Unique identifier for a chat.
      example: 1
    chat_company:
      type: string
      description: username of the company associated with a chat.
      example: companyA
    unread:
      type: integer
      description: Number of unread messages in the chat.
      example: 5
    Chat:
      type: object
      properties:
        chat_id:
          $ref: "#/components/schemas/chat_id"
        company:
          $ref: "#/components/schemas/chat_company"
        unread:
          $ref: "#/components/schemas/unread"
      required:
        - chat_id
        - company
        - unread
      description: A chat belonging to a company
...
    container_id:
      type: integer
      description: Unique identifier for the container.
      example: 101
    max_weight:
      type: number
      description: Maximum weight capacity of the container in tons.
      example: 100
    Container:
      type: object
      properties:
        container_id:
          $ref: "#/components/schemas/container_id"
        max_weight:
          $ref: "#/components/schemas/max_weight"
      required:
        - container_id
        - max_weight
      description: A shipping container
    ContainerCreate:
      type: object
      properties:
        container_id:
          $ref: "#/components/schemas/container_id"
        max_weight:
          $ref: "#/components/schemas/max_weight"
      required:
        - max_weight
      description: Data transfer object for creating a new Container.
      x-fiddle-dto-info:
        baseSchemaName: Container
...
    Order:
      type: object
      properties:
        order_id:
          type: integer
          description: Unique identifier of the order.
          example: 123
        requester:
          type: string
          description: User who requested the order.
          example: companyA
        made_at:
          type: string
          format: date-time
          example: 2024-06-01T08:00:00Z
        source:
          type: string
          description: Username associated with the starting location of the order.
          example: companyA
        destination:
          type: string
          description: Username associated with the destination of the order.
          example: companyA
        route_id:
          type: integer
          description: ID of the route associated with the order.
          example: 12
        tw_start:
          type: string
          format: date-time
          example: 2024-06-01T08:00:00Z
        tw_end:
          type: string
          format: date-time
          example: 2024-06-01T08:00:00Z
        status:
          type: string
          enum:
            - created
            - scheduled
            - undergoing
            - completed
            - canceled
          example: created
        weight:
          type: number
          example: 12
        operation_type:
          type: string
          enum:
            - loading
            - unloading
          example: loading
        freight_value:
          type: number
          description: value of the contents in EUR.
          example: 12
        is_shared:
          type: boolean
        container_id:
          type: integer
        eta:
          type: string
          format: date-time
          example: 2024-06-01T08:00:00Z
      required:
        - order_id
        - requester
        - made_at
        - source
        - destination
        - route_id
        - tw_start
        - tw_end
        - status
        - weight
        - operation_type
        - freight_value
        - is_shared
        - container_id
        - eta
      description: An order made by a company
...
    Company:
      type: object
      properties:
        username:
          type: string
          example: companyA
        password:
          type: string
          example: 4$zKbc9GZnZ1UOeQ7IgGDsDQ$Oq3U6n+cNmC8MUt1kUZ1E/cNaJZ6FS69aUlFoCqZLXE
        name:
          type: string
          example: A
        location:
          type: object
          properties:
            latitude:
              type: number
              example: 41.1
            longitude:
              type: number
              example: 41.1
          required:
            - latitude
            - longitude
        closing_time:
          type: string
          format: time
          example: 14:30:00
        opening_time:
          type: string
          format: time
          example: 14:30:00
      required:
        - username
        - password
        - name
        - location
        - closing_time
        - opening_time
      description: A simplified representation of a Company, typically used in list views.
    CompanySummary:
      type: object
      properties:
        username:
          type: string
          example: companyA
        name:
          type: string
          example: A
        location:
          type: object
          properties:
            latitude:
              type: number
              example: 41.1
            longitude:
              type: number
              example: 41.1
          required:
            - latitude
            - longitude
        closing_time:
          type: string
          format: time
          example: 14:30:00
        opening_time:
          type: string
          format: time
          example: 14:30:00
      required:
        - username
        - name
        - location
        - closing_time
        - opening_time
      description: A simplified representation of a Company.
      x-fiddle-dto-info:
        baseSchemaName: Company
    CompanyPartialUpdate:
      type: object
      properties:
        username:
          type: string
          example: companyA
        password:
          type: string
        name:
          type: string
          example: A
        location:
          type: object
          properties:
            latitude:
              type: number
              example: 41.1
            longitude:
              type: number
              example: 41.1
          required:
            - latitude
            - longitude
        closing_time:
          type: string
          format: time
          example: 14:30:00
        opening_time:
          type: string
          format: time
          example: 14:30:00
      required: []
      description: Data transfer object for partially updating an existing Company
        (PATCH operation).
      x-fiddle-dto-info:
        baseSchemaName: Company
...